Refactor(Plugins): Fix an issue with plenary and luarocks
Lazy for some reason was unable to download luarocks during the plenary build. so plenary was not built, and this plugin is used as dependency in many other plugins. so basically I need it. it dosent need me, but I do.
I just disabled rocks in the following file:
lua\asthriona\plugins\init.lua